I can put objects into the world. I also redesigned how config works so there's no less string comparisons. I also fucked around with graphics because the next top priority is splitting view and model up

+++ src/animated_sprite.hpp
@@ -18,7 +18,7 @@ public:
 
     void play_animation(const std::string &animation, bool once = false);
 
-    void update(int elapsed_time);
+    void update(float elapsed_time);
 
     void draw(Graphics &graphics, int x, int y);
